abstracting storyboard related stuff
