+++ BDDComicsList/Base.lproj/Main.storyboard
@@ -13,7 +13,7 @@
         <!--Marvel Comics-->
         <scene sceneID="tne-QT-ifu">
             <objects>
-                <viewController id="BYZ-38-t0r" customClass="ComicsListViewController" customModule="BDDComicsList" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="ComicsListViewController" id="BYZ-38-t0r" customClass="ComicsListViewController" customModule="BDDComicsList" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="8bC-Xf-vdC">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
